requirements.txt
# Core dependencies for TranscribeTalk
openai-whisper==20231117
openai==1.12.0
elevenlabs==0.2.26
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pydantic==2.5.3
aiofiles==23.2.1

# Development dependencies
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0 